+++ packages/app-andrew/src/pages/group/[...params].tsx
@@ -15,6 +15,7 @@ import { IPost } from '@gtms/commons/models'
 // api
 import { fetchPost } from '@gtms/api-post'
 import { findTagsAPI, fetchSuggestedTagsAPI } from '@gtms/api-tags'
+import { findbyUsernameAPI } from '@gtms/api-auth'
 // components
 import { FavsButton } from 'components/group/FavsButton'
 import { FollowButton } from 'components/group/FollowButton'
@@ -247,6 +248,7 @@ const GroupPage: NextPage<GroupPageProps> = (props) => {
                   </NavigationTabs>
                   <PostCreate
                     fetchTags={findTagsAPI}
+                    fetchUsers={findbyUsernameAPI}
                     fetchSuggestedTags={fetchSuggestedTagsAPI}
                     user={state.user}
                     noImage={UserAvatarNoImage}
@@ -261,6 +263,7 @@ const GroupPage: NextPage<GroupPageProps> = (props) => {
                   />
                   <RecentlyAddedPosts
                     fetchTags={findTagsAPI}
+                    fetchUsers={findbyUsernameAPI}
                     onPostClick={onPostClick}
                     onTagClick={(tag) => {
                       router.push(`/group/${state.group?.slug}/tag/${tag}`)