+++ packages/state-post/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.3](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.8.2...@gtms/state-post@0.8.3) (2020-07-19)
+
+**Note:** Version bump only for package @gtms/state-post
+
+
+
+
+
+## [0.8.2](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.8.1...@gtms/state-post@0.8.2) (2020-07-19)
+
+**Note:** Version bump only for package @gtms/state-post
+
+
+
+
+
 ## [0.8.1](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.8.0...@gtms/state-post@0.8.1) (2020-07-15)
 
 **Note:** Version bump only for package @gtms/state-post